Split type hints for Quantity into .compat.xarray
